Cambio para que el tipo decimal de .net devuelva como decimal al buscar en el diccionario para generar los modelos y dtos

+++ Capibara/Capas.cs
@@ -18,7 +18,7 @@ namespace Capibara
             {typeof(char[]),    ERROR},
             {typeof(DateTime),  "DateTime"},
             {typeof(DBNull),    ERROR},
-            {typeof(decimal),   "double"},
+            {typeof(decimal),   "decimal"},
             {typeof(double),    "double"},
             {typeof(Guid),      "Guid"},
             {typeof(int),       "int"},			//                                                                      Int32
